Day 4: Class vs. Instance     https://www.hackerrank.com/challenges/30-class-vs-instance/problem
if age <13, print You are young
if age >=13&<18, print You are a teenager.
Otherwise, print You are old..

4
-1
10
16
18

Age is not valid, setting age to 0.
You are young.
You are young.

You are young.
You are a teenager.

You are a teenager.
You are old.

You are old.
You are old.